feat(api): remove HTTP service
