+++ src/shared/api/index.ts
@@ -2,29 +2,99 @@ import type {
   Account,
   AssetDefinition,
   Asset,
-  AssetSearchParams,
-  Block,
   Domain,
-  Paginated,
   PaginationParams,
+  Block,
+  Paginated,
   TransactionSearchParams,
+  AssetSearchParams,
   Transaction,
   PeerStatus,
-} from '@/shared/api/dto';
+  AccountId,
+  AssetId,
+  AssetDefinitionId,
+  DomainId,
+} from './schemas';
+import {
+  peerStatusSchema,
+  accountSchema,
+  assetSchema,
+  assetDefinitionSchema,
+  domainSchema,
+  blockSchema,
+  paginatedSchema,
+  transactionSchema,
+} from './schemas';
 
-interface HTTPService {
-  fetchAccounts: (params?: PaginationParams) => Promise<Paginated<Account>>
-  fetchAccount: (id: string) => Promise<Account>
-  fetchAssets: (params?: AssetSearchParams) => Promise<Paginated<Asset>>
-  fetchAsset: (id: string) => Promise<Asset>
-  fetchAssetDefinitions: (params?: PaginationParams) => Promise<Paginated<AssetDefinition>>
-  fetchAssetDefinition: (id: string) => Promise<AssetDefinition>
-  fetchDomains: (params?: PaginationParams) => Promise<Paginated<Domain>>
-  fetchDomain: (id: string) => Promise<Domain>
-  fetchBlocks: (params?: PaginationParams) => Promise<Paginated<Block>>
-  fetchBlock: (heightOrHash: number | string) => Promise<Block>
-  fetchTransactions: (params?: TransactionSearchParams) => Promise<Paginated<Transaction>>
-  fetchPeerStatus: () => Promise<PeerStatus>
+const BASE_URL = import.meta.env.VITE_API_URL ?? '/api/v1';
+
+async function get<T>(path: string, params?: PaginationParams): Promise<T> {
+  const url = new URL(`${BASE_URL}${path}`);
+
+  if (params) {
+    Object.entries(params).forEach(([key, value]) => url.searchParams.set(key, String(value)));
+  }
+
+  const res = await fetch(url);
+  return res.json();
 }
 
-export const http: HTTPService = await import('./http');
+export async function fetchAccounts(params?: PaginationParams): Promise<Paginated<Account>> {
+  const res = await get('/accounts', params);
+  return paginatedSchema(accountSchema).parse(res);
+}
+
+export async function fetchAccount(id: AccountId): Promise<Account> {
+  const res = await get(`/accounts/${id.toString()}`);
+  return accountSchema.parse(res);
+}
+
+export async function fetchAssets(params?: AssetSearchParams): Promise<Paginated<Asset>> {
+  const res = await get('/assets', params);
+  return paginatedSchema(assetSchema).parse(res);
+}
+
+export async function fetchAsset(id: AssetId): Promise<Asset> {
+  const res = await get(`/assets/${id.toString()}`);
+  return assetSchema.parse(res);
+}
+
+export async function fetchAssetDefinitions(params?: PaginationParams): Promise<Paginated<AssetDefinition>> {
+  const res = await get('/assets-definitions', params);
+  return paginatedSchema(assetDefinitionSchema).parse(res);
+}
+
+export async function fetchAssetDefinition(id: AssetDefinitionId): Promise<AssetDefinition> {
+  const res = await get(`/assets-definitions/${encodeURIComponent(id.toString())}`);
+  return assetDefinitionSchema.parse(res);
+}
+
+export async function fetchDomains(params?: PaginationParams): Promise<Paginated<Domain>> {
+  const res = await get('/domains', params);
+  return paginatedSchema(domainSchema).parse(res);
+}
+
+export async function fetchDomain(id: DomainId): Promise<Domain> {
+  const res = await get(`/domains/${id.getDomain()}`);
+  return domainSchema.parse(res);
+}
+
+export async function fetchBlocks(params?: PaginationParams): Promise<Paginated<Block>> {
+  const res = await get('/blocks', params);
+  return paginatedSchema(blockSchema).parse(res);
+}
+
+export async function fetchBlock(heightOrHash: number | string): Promise<Block> {
+  const res = await get(`/blocks/${heightOrHash}`);
+  return blockSchema.parse(res);
+}
+
+export async function fetchPeerStatus(): Promise<PeerStatus> {
+  const res = await get('/status');
+  return peerStatusSchema.parse(res);
+}
+
+export async function fetchTransactions(params?: TransactionSearchParams): Promise<Paginated<Transaction>> {
+  const res = await get('/transactions', params);
+  return paginatedSchema(transactionSchema).parse(res);
+}
